refactor: withdraw NavItem 추가
