+++ src/view/group/index/Withdraw/index.tsx
@@ -1,7 +1,7 @@
 import Button from "@/common/component/Button";
-import SuccessWithdraw from "@/view/group/index/component/Withdraw/SuccessWithdraw";
-import PromptWithdraw from "@/view/group/index/component/Withdraw/\bPromptWithdraw";
-import { withdrawWrapper } from "@/view/group/index/component/Withdraw/index.css";
+import PromptWithdraw from "@/view/group/index/Withdraw/PromptWithdraw";
+import SuccessWithdraw from "@/view/group/index/Withdraw/SuccessWithdraw";
+import { withdrawWrapper } from "@/view/group/index/Withdraw/index.css";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
